// Web/src/generated/model/auditFilterOptionField.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReleaseHub API
 * Shared HTTP API contract for the ReleaseHub frontend and backend.
 * OpenAPI spec version: 1.0.0
 */

export type AuditFilterOptionField = typeof AuditFilterOptionField[keyof typeof AuditFilterOptionField];


export const AuditFilterOptionField = {
  action: 'action',
  resourceType: 'resourceType',
  actor: 'actor',
} as const;